requirements.txt
numpy>=1.24
torch>=2.0
web3>=6.11
py-solc-x>=1.1.1
pandas>=2.0
scikit-learn>=1.3
matplotlib>=3.7
